Refactored conftest.py
This closes #86
